+++ src/Widget/CheckboxList/CheckboxList.php
@@ -109,9 +109,8 @@ final class CheckboxList implements NoEncodeStringableInterface
      */
     public function value(...$value): self
     {
-        /** @psalm-var list<scalar|\Stringable> $value */
         $new = clone $this;
-        $new->values = array_map('\strval', $value);
+        $new->values = array_map('\strval', array_values($value));
         return $new;
     }