+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Bug #68: incorrect work of `TagContentTrait::content` and `TagContentTrait::addContent` methods when used named parameters (vjik)
 
 ## 1.0.0 March 17, 2021